+++ src/2_mlx_setup/setup_mlx.c
@@ -6,7 +6,7 @@
 /*   By: freddy <freddy@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/28 08:41:35 by freddy            #+#    #+#             */
-/*   Updated: 2024/05/28 09:50:10 by freddy           ###   ########.fr       */
+/*   Updated: 2024/05/28 10:29:36 by freddy           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -49,4 +49,6 @@ void	setup_mlx(void)
 	data = get_persistent_data();
 	data->mlx = mlx;
 	set_sky();
+	mlx_loop_hook(data->mlx, loop_hook, NULL);
+	mlx_key_hook(data->mlx, key_hook, NULL);
 }