+++ CHANGELOG.md
@@ -1,11 +1,15 @@
 # Changelog
 
-### [0.7.1] - 28.05.2024 - Freddy
+### [0.7.2] - 28.05.2024 - Freddy - `14-very-basic-rendering-setup`
+- added hook setup & loop hook & key hook
+- made program close when esc is received
+
+### [0.7.1] - 28.05.2024 - Freddy - `14-very-basic-rendering-setup`
 - Made t_color struct hold unsigned chars, more efficient.
 - Made colors as ints use unsigned int instead, as that keeps the data properly
 - ceiling & floor colors now render properly
 
-## [0.7.0] - 28.05.2024 - Freddy
+## [0.7.0] - 28.05.2024 - Freddy - `14-very-basic-rendering-setup`
 - Added an alpha field to color struct & added conversion to and from int utils
 - added mlx setup function & created basic window & drew background white. currently the colors dont appear properly.